experiments/veo-app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
absl-py==2.3.1
    # via mesop
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
asttokens==3.0.0
    # via stack-data
black==25.1.0
    # via veo-app
blinker==1.9.0
    # via flask
cachecontrol==0.14.3
    # via firebase-admin
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   flask
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   ipython
    #   pytest
contourpy==1.3.3
    # via matplotlib
cryptography==45.0.5
    # via pyjwt
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
deepdiff==6.7.1
    # via mesop
docstring-parser==0.17.0
    # via google-cloud-aiplatform
executing==2.2.0
    # via stack-data
fastapi==0.116.1
    # via veo-app
firebase-admin==6.9.0
    # via veo-app
flask==3.1.1
    # via mesop
fonttools==4.59.0
    # via matplotlib
google-api-core==2.25.1
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-api-python-client==2.177.0
    # via firebase-admin
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform==1.105.0
    # via veo-app
google-cloud-bigquery==3.35.1
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.21.0 ; platform_python_implementation != 'PyPy'
    # via firebase-admin
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via
    #   firebase-admin
    #   google-cloud-aiplatform
    #   veo-app
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.20.0
    # via
    #   google-cloud-aiplatform
    #   veo-app
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
gunicorn==23.0.0
    # via veo-app
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   firebase-admin
    #   google-genai
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
ipython==9.4.0
    # via mediapy
ipython-pygments-lexers==1.1.1
    # via ipython
itsdangerous==2.2.0
    # via flask
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via flask
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via mediapy
matplotlib-inline==0.1.7
    # via ipython
mediapy==1.2.4
    # via veo-app
mesop==1.1.0
    # via veo-app
msgpack==1.1.1
    # via
    #   cachecontrol
    #   mesop
mypy-extensions==1.1.0
    # via black
numpy==2.3.2
    # via
    #   contourpy
    #   matplotlib
    #   mediapy
    #   pandas
    #   shapely
    #   veo-app
ordered-set==4.1.0
    # via deepdiff
packaging==25.0
    # via
    #   black
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   gunicorn
    #   matplotlib
    #   pytest
pandas==2.3.1
    # via veo-app
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via ipython
pillow==11.3.0
    # via
    #   matplotlib
    #   mediapy
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.51
    # via ipython
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   mesop
    #   proto-plus
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22 ; platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   mesop
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pyjwt==2.10.1
    # via firebase-admin
pyparsing==3.2.3
    # via
    #   httplib2
    #   matplotlib
pytest==8.4.1
    # via veo-app
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via mesop
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
ruff==0.12.7
    # via veo-app
shapely==2.1.1
    # via google-cloud-aiplatform
shortuuid==1.0.13
    # via veo-app
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
stack-data==0.6.3
    # via ipython
starlette==0.47.2
    # via fastapi
tenacity==9.1.2
    # via veo-app
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.14.1
    # via
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via veo-app
watchdog==6.0.0
    # via mesop
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via
    #   flask
    #   mesop
